Added the keypad control: now inserts successfully the numbers. Corrected the KeypadNumber class and also the CommandCreator function fillNchars

+++ KeypadNumber.h
@@ -15,8 +15,8 @@
       int getLength();
   
     private:
-      byte _digits[NDIGITS];
-      byte _index;
+      int _digits[NDIGITS];
+      int _index;
   };
 
 #endif
